feat : Simplifying smaple app
