--- sample/src/main/java/org/aerogear/graphqlandroid/MyApplication.kt
@@ -1,22 +1,0 @@
-package org.aerogear.graphqlandroid
-
-import android.app.Application
-import android.arch.persistence.room.Room
-import android.util.Log
-import org.aerogear.graphqlandroid.persistence.Database
-
-class MyApplciation : Application() {
-
-    companion object {
-        lateinit var database: Database
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        database = Room.databaseBuilder(this, Database::class.java, "roomDb.db")
-            .fallbackToDestructiveMigration()
-            .allowMainThreadQueries()
-            .build()
-        Log.e("TAG", "Application was started")
-    }
-}
