--- sample/src/main/java/org/aerogear/graphqlandroid/persistence/Database.kt
@@ -1,13 +1,0 @@
-package org.aerogear.graphqlandroid.persistence
-
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
-import android.arch.persistence.room.TypeConverters
-
-@Database(entities = arrayOf(Mutation::class), version = 1)
-@TypeConverters(Converters::class)
-abstract class Database : RoomDatabase() {
-
-    abstract fun mutationDao(): MutationDao
-
-}
