+++ sample/src/main/res/layout/activity_main.xml
@@ -1,10 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout
+
+<LinearLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
+        android:orientation="vertical"
         android:layout_height="match_parent">
 
+    <LinearLayout android:layout_width="match_parent"
+                  android:layout_height="wrap_content">
+
+        <Button android:layout_width="0dp"
+                android:layout_weight="1"
+                android:id="@+id/insertbutton"
+                android:layout_margin="16dp"
+                android:checked="true"
+                android:text="Insert"
+                android:background="#008577"
+                android:textColor="#000000"
+                android:layout_height="44dp"/>
+
+
+        <Button android:layout_width="0dp"
+                android:layout_weight="1"
+                android:id="@+id/deletetask"
+                android:checked="true"
+                android:text="Delete"
+                android:background="#008577"
+                android:layout_margin="16dp"
+                android:textColor="#000000"
+                android:layout_height="44dp"/>
+
+
+    </LinearLayout>
 
     <android.support.v4.widget.SwipeRefreshLayout
             android:layout_width="match_parent"
@@ -14,6 +42,7 @@
 
         <android.support.v7.widget.RecyclerView
                 android:layout_margin="4dp"
+                android:layout_marginLeft="8dp"
                 android:id="@+id/recycler_view"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"/>
@@ -22,13 +51,5 @@
     </android.support.v4.widget.SwipeRefreshLayout>
 
 
-    <android.support.design.widget.FloatingActionButton
-            android:id="@+id/fabAdd"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="end|bottom"
-            android:layout_margin="32dp"
-            app:srcCompat="@drawable/add"/>
 
-
-</android.support.design.widget.CoordinatorLayout>
+</LinearLayout>
